subtracting from total working, updating component not working

+++ db/migrate/20160401204237_create_bills.rb
@@ -2,7 +2,7 @@ class CreateBills < ActiveRecord::Migration
   def change
     create_table :bills do |t|
       t.string :name , null: false
-      t.integer :total, null: false
+      t.decimal :total, null: false
       t.belongs_to :money
       t.timestamps null: false
     end
